# AMOS_OPENAPI_SPEC.yaml
openapi: 3.1.0
info:
  title: AMOS Platform API
  description: |
    Unified API Gateway for the AMOS 5-Repository Integration Architecture.
    
    This API serves as the central communication hub for:
    - AMOS-Code (core brain package)
    - AMOS-Consulting (this service - platform hub)
    - AMOS-Claws (operator frontend)
    - Mailinhconect (product frontend)
    - AMOS-Invest (investor dashboard)
    
    Communication lanes:
    - Sync: REST API + WebSocket
    - Async: Redis Streams / NATS (event bus)
  version: 1.0.0
  contact:
    name: AMOS Platform Team
    email: platform@amos.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.amos.io/v1
    description: Production
  - url: https://api-staging.amos.io/v1
    description: Staging
  - url: http://localhost:8000/v1
    description: Local Development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /chat:
    post:
      summary: Send Chat Message
      description: Send a message and get AI response
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'

  # Agent Operations
...
  /repo/scan:
    post:
      summary: Scan Repository
      description: Initiate a repository scan for issues
      tags:
        - Repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepoScanRequest'
      responses:
        '202':
          description: Scan initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanAccepted'
  /repo/fix:
    post:
      summary: Fix Repository
      description: Apply automated fixes to a repository
      tags:
        - Repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepoFixRequest'
      responses:
        '202':
          description: Fix job initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FixAccepted'
...
  /workflow/run:
    post:
      summary: Run Workflow
      description: Execute a workflow
      tags:
        - Workflows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowRunRequest'
      responses:
        '202':
          description: Workflow started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowAccepted'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  schemas:
...
    ChatRequest:
      type: object
      required:
        - message
        - workspace_id
      properties:
        message:
          type: string
          description: User message
        context:
          type: array
          items:
            $ref: '#/components/schemas/Message'
          description: Previous messages for context
        model:
          type: string
          description: Model to use (default from config)
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 0.7
        max_tokens:
          type: integer
          minimum: 1
          maximum: 8192
        workspace_id:
          type: string
          description: Workspace/tenant ID
    Message:
      type: object
      properties:
        role:
          type: string
          enum: [system, user, assistant]
        content:
          type: string
        timestamp:
          type: string
          format: date-time
    ChatResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        message:
          type: string
        model:
          type: string
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
        timestamp:
          type: string
          format: date-time
...
    RepoScanRequest:
      type: object
      required:
        - repo_url
      properties:
        repo_url:
          type: string
          format: uri
        branch:
          type: string
          default: main
        scan_types:
          type: array
          items:
            type: string
            enum: [security, performance, style, architecture, dependencies, tests]
        depth:
          type: string
          enum: [quick, standard, deep]
          default: standard
...
    RepoFixRequest:
      type: object
      required:
        - repo_url
        - scan_id
      properties:
        repo_url:
          type: string
        scan_id:
          type: string
        fix_types:
          type: array
          items:
            type: string
        create_pr:
          type: boolean
          default: true
    ScanAccepted:
      type: object
      properties:
        scan_id:
          type: string
        status:
          type: string
          enum: [queued, running]
    FixAccepted:
      type: object
      properties:
        fix_id:
          type: string
        status:
          type: string
    WorkflowRunRequest:
      type: object
      required:
        - workflow_id
      properties:
        workflow_id:
          type: string
        parameters:
          type: object
          additionalProperties: true
        trigger:
          type: string
          enum: [manual, scheduled, webhook, event]
    WorkflowAccepted:
      type: object
      properties:
        workflow_id:
          type: string
        execution_id:
          type: string
        status:
          type: string
...
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
            request_id:
              type: string
            timestamp:
              type: string
              format: date-time
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...